AvilaOps/data/dashboards/executive/requirements.txt
# Ávila Executive Dashboard System - Requirements
# ============================================

# Core Dependencies
plotly>=5.17.0
pandas>=2.1.0
numpy>=1.24.0

# Email & SMTP
smtplib3>=0.1.1
email-validator>=2.1.0

# Async & Scheduling
asyncio>=3.4.3
schedule>=1.2.0
aiofiles>=23.2.1

# Data Processing
openpyxl>=3.1.2
python-dateutil>=2.8.2

# Logging & Configuration
colorama>=0.4.6
python-dotenv>=1.0.0

# Web & Templates
Jinja2>=3.1.2
beautifulsoup4>=4.12.2

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
black>=23.7.0
flake8>=6.0.0

# Optional: Enhanced Features
# ---------------------------
# For future web dashboard
# fastapi>=0.103.0
# uvicorn>=0.23.0

# For advanced analytics
# scikit-learn>=1.3.0
# matplotlib>=3.7.0
# seaborn>=0.12.0

# For database integration
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0

# For cloud deployment
# boto3>=1.28.0
# azure-storage-blob>=12.17.0